+++ kubejs/client_scripts/conflicts.js
@@ -1,0 +1,3 @@
+JEIEvents.hideItems(event => {
+    event.hide("vinery:coarse_dirt_slab");
+});