+++ kubejs/server_scripts/conflicts.js
@@ -1,0 +1,3 @@
+ServerEvents.recipes((event) => {
+    event.remove({ id: "vinery:coarse_dirt_slab" });
+});